BUGFIX: Fix baseUri handling for with Neos > 3.3
In Neos 3.3 the Uri object implements the PSR7 Uri interface wich enforces that the getHost method always returns a string. 
Since the existing implementation of getLinkValue explicitly checked for null currently the detection of host-relative urls did not work any more wich resulted in loosing the baseUri for linked resources.

This change check for the host beeing null or an empty string and thus supports old and new Neos versions.